JS File management
Dividing js functions in separate files
