+++ public/js/slider/arrows.js
@@ -1,0 +1,14 @@
+import { goLeft, goRight } from "./images.js";
+
+// Arrows fn
+
+/** Creates the slider left amd right arrows & adds the click event listener */
+
+function makeArrows() {
+  const leftArrow = document.getElementById("arrowLeft");
+  const rightArrow = document.getElementById("arrowRight");
+  rightArrow.addEventListener("click", goRight);
+  leftArrow.addEventListener("click", goLeft);
+}
+
+export { makeArrows };
